+++ domain/src/main/java/com/london/domain/repository/AddMovieRatingRepository.kt
@@ -3,6 +3,6 @@ package com.london.domain.repository
 interface AddMovieRatingRepository {
     suspend fun addMovieRatingById(
         id: Int,
-        rating: Double,
+        rating: Int,
     ) : Boolean
 }
